v3

+++ Homework 6.1.py
@@ -5,7 +5,7 @@ with open('departments.json', 'r') as f:
 for departments in dep['departments']:
     if departments['expenses'] < departments['income']:
         for employee in departments['employees']:
-            employee['salary'] = employee['salary'] * 1.1
+            employee['salary'] = int(employee['salary'] * 1.1)
 
 with open('new_costs.json', 'w') as f:
      f.write(json.dumps(dep, indent=4))
